Update README: TOC, .NET 4.8, Usage, License
Added a Table of Contents section to the README.md file for better navigation. Updated the .NET Framework requirement from version 4.5.2 to 4.8. Improved the installation instructions by clarifying the paths for the MusicBee Plugins directory. Added a new Usage section to guide users on how to use the plugin after installation. Added a License section to specify that the project is licensed under the MIT License and provided a link to the LICENSE file.